refactor: more refactoring (#93)
