+++ components/Board/localSelectSquare.js
@@ -1,6 +1,6 @@
 import { getNumberFromLetter } from "../Board/board.js";
 import { LETTERS } from "../Board/board.js";
-import King from "../../pieces/King.js";
+import King from "../../services/pieces/King.js";
 import convertToChessNotation from "../Logger/toChessNotation.js";
 import { EmitActions } from "../../constants/Actions.js";
 import {
