+++ screens/ProfileScreen/GameStat.jsx
@@ -1,10 +1,10 @@
 import { useRoute } from "@react-navigation/native";
 import { View, Text } from "react-native";
-import axiosInstance from "../../components/axiosInstance";
 import React, { useState, useEffect } from "react";
 import ArchivedGamesContainer from "./ArchivedGamesContainer";
 import styles from "./gameStatStyles";
 import * as SecureStore from "expo-secure-store";
+import axiosInstance from "@/services/axios/axiosInstance";
 
 const GameStat = ({ navigation }) => {
   const route = useRoute();
