+++ services/socket/socketHandler.js
@@ -1,4 +1,4 @@
-import Actions from "../constants/Actions";
+import Actions from "../../constants/Actions";
 import {
   handleGameStart,
   handleMoveMade,
